Drop nbxplorer and bitcoind from integration tests setup
